pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "atencion-cliente-bot"
version = "0.1.0"
description = "Bot profesional de atención al cliente"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Team", email = "dev@example.com"}]
license = { file = "LICENSE" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
]
dependencies = [
  "python-telegram-bot>=20.7",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.22.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0",
]

[tool.setuptools]
packages = ["src"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
pythonpath = ["src"]
